+++ graphics/include/gz/common/Image.hh
@@ -110,8 +110,8 @@ namespace gz
       /// \param[in] _filename The name of the saved image
       public: void SavePNG(const std::string &_filename);
 
-      /// \brief Save the image in PNG format
-      /// \param[in] _filename The name of the saved image
+      /// \brief Get the PNG image in a buffer
+      /// \param[out] _buffer Buffer with the data
       public: void SavePNGToBuffer(std::vector<unsigned char> &_buffer);
 
       /// \brief Set the image from raw data